Portfolio Version 1 - Dynamic/Portfolio/Backend/requirements.txt
 
fastapi
uvicorn
google-generativeai
pydantic
